backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pandas
googlemaps
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
